:white_check_mark: Add flake8 configuration and refactor to fix flake8 errors
