+++ octopusapi/__init__.py
@@ -1,3 +1,10 @@
 __version__ = "0.1.0"
 from .electricity_meter import ElectricityMeter
 from .exceptions import OctopusException, ApiException, ParameterException
+
+__all__ = [
+    "ApiException",
+    "ElectricityMeter",
+    "OctopusException",
+    "ParameterException",
+]
