+++ tests/text_fixtures.py
@@ -1,0 +1,10 @@
+def test_have_api_key(api_key):
+    assert api_key is not None
+
+
+def test_have_electricity_mpan(electricity_mpan):
+    assert electricity_mpan is not None
+
+
+def test_have_electricity_serial_number(electricity_serial_number):
+    assert electricity_serial_number is not None
